% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ensembl_to_hgnc.R
\name{convert_ensembl_to_hgnc}
\alias{convert_ensembl_to_hgnc}
\title{Ensembl to HGNC Symbol}
\usage{
convert_ensembl_to_hgnc(ensembl_id)
}
\arguments{
\item{ensembl_id}{ensembl id (character)}
}
\value{
HGNC approved symbols (character). NA if can't find a match.
}
\description{
Converts ensembl ids to HGNC approved symbol.
}
