% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hgnc_to_ensembl.R
\name{convert_hgnc_to_ensembl}
\alias{convert_hgnc_to_ensembl}
\title{Mapping HGNC to ensembl ids}
\usage{
convert_hgnc_to_ensembl(hgnc_symbols)
}
\arguments{
\item{hgnc_symbols}{HGNC symbols. Can be the currently approved symbol, an alias, or a previous symbols (character or type that can be converted to a character)}
}
\value{
ensembl id (string)
}
\description{
Maps HGNC gene symbols to ensembl_ids
}
