% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ensembl_to_hgnc.R
\name{convert_single_ensembl_to_hgnc}
\alias{convert_single_ensembl_to_hgnc}
\title{Ensembl to HGNC Symbol}
\usage{
convert_single_ensembl_to_hgnc(ensembl_id)
}
\arguments{
\item{ensembl_id}{ensembl id (string)}
}
\value{
HGNC approved symbols (string). NA if can't find a match.
}
\description{
Converts a single ensembl id to HGNC approved symbol.
}
