% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_hgnc_database.R
\name{preprocess_hgnc_database}
\alias{preprocess_hgnc_database}
\title{preprocess_hgnc_database}
\usage{
preprocess_hgnc_database()
}
\value{
NULL. Run for its side effects
}
\description{
Takes the HGNC database from:https://biomart.genenames.org/ and adds turns it into a redundant longform dataset with a HGNC.SYMBOL column that represents each and every HGNC symbol (approved, alternative, and previous)
}
\examples{
preprocess_hgnc_database()
}
